+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog for PYDIFACT
 - Deprecate SegmentCollection
 - Add Interchange and Message semantic and featured classes
 - Add an optional predicate filter to `get_segment` and `get_segments` methods
+- Allow to access AbstractSegmentsContainer elements using `container[i]`
 
 [0.0.5] - 2020-04-29
 - corrected pypi details